transition alpha weirdness
this isn't a perfect solution but at least it looks a lot better.
The problem happens when the 'top' layer has a layer underneath
that has a different color thant he rest of the other things that
look like a top layer.  For example, the door archway has a patch
of black shadow underneath which make it alpha blend differently
than the rest of it's neighbors.
I guess the way to 'fix' this is to alpha blend only what's visible
and not the actual tiles but that's out of scope right now.
The quick and dirty fix is to not render tiles in the last column
of the dungeon tileset (which are the shadows of doorways) and place
another purple tile under the archway.  Where the shadow is still
flashes in transition but this is much less noticeable than the
archway not being properly alpha blended.